refactor: 응답 DTO의 패키지 위치를 역할에 맞게 변경
