+++ backend/src/main/java/com/backend/petplace/domain/point/controller/PointController.java
@@ -1,7 +1,7 @@
 package com.backend.petplace.domain.point.controller;
 
+import com.backend.petplace.domain.point.dto.response.PointHistoryResponse;
 import com.backend.petplace.domain.point.service.PointService;
-import com.backend.petplace.domain.review.dto.response.PointHistoryResponse;
 import com.backend.petplace.global.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
