--- backend/src/main/java/com/backend/petplace/domain/point/dto/request/PointRequest.java
@@ -1,5 +1,0 @@
-package com.backend.petplace.domain.point.dto.request;
-
-public class PointRequest {
-
-}
