+++ backend/src/main/java/com/backend/petplace/domain/point/dto/response/PointHistoryResponse.java
@@ -1,4 +1,4 @@
-package com.backend.petplace.domain.review.dto.response;
+package com.backend.petplace.domain.point.dto.response;
 
 import com.backend.petplace.domain.review.dto.PointTransaction;
 import io.swagger.v3.oas.annotations.media.Schema;
