--- backend/src/main/java/com/backend/petplace/domain/point/dto/response/PointResponse.java
@@ -1,5 +1,0 @@
-package com.backend.petplace.domain.point.dto.response;
-
-public class PointResponse {
-
-}
