+++ backend/src/main/java/com/backend/petplace/domain/review/controller/ReviewController.java
@@ -3,7 +3,6 @@ package com.backend.petplace.domain.review.controller;
 import com.backend.petplace.domain.review.dto.request.ReviewCreateRequest;
 import com.backend.petplace.domain.review.dto.response.MyReviewResponse;
 import com.backend.petplace.domain.review.dto.response.PlaceReviewsResponse;
-import com.backend.petplace.domain.review.dto.response.PointHistoryResponse;
 import com.backend.petplace.domain.review.dto.response.ReviewCreateResponse;
 import com.backend.petplace.domain.review.service.ReviewService;
 import com.backend.petplace.global.response.ApiResponse;
